Added Textures and Upgrades
added publically available prefabs for projectiles
added upgrade mechanism to upgrade the player/projectiles upon rarity
and i dont rememebr what all i did cos i have been making changed continuosly for past 2 days and never committed the changes :)

+++ Assets/Interface/IHealth.cs
@@ -8,7 +8,7 @@ namespace Assets.Interface
 {
     public interface IHealth
     {
-        int Health { get; set; }
+        double Health { get; set; }
         bool IsDead => Health <= 0;
     }
 }
